fixed up most of the storage system ui base

+++ src/main/java/com/github/klikli_dev/occultism/client/gui/controls/GuiMachineSlot.java
@@ -72,7 +72,7 @@ public class GuiMachineSlot {
     //endregion Getter / Setter
 
     //region Methods
-    public boolean isMouseOverSlot(int mouseX, int mouseY) {
+    public boolean isMouseOverSlot(double mouseX, double mouseY) {
         return this.parent.isPointInRegion(this.x - this.guiLeft, this.y - this.guiTop, 16, 16, mouseX, mouseY);
     }
